mpd check bugfix

+++ Nix-Relic/modules/system/services/mpdchck.nix
@@ -28,54 +28,58 @@ with lib; let
         pw-metadata -n settings 0 clock.force-rate 0 > /dev/null
 
         break
-      else
-        currentRate=$(mpc status %samplerate%)
+      fi
+      currentRate=$(mpc status %samplerate%)
+      if [ "$currentRate" = "%samplerate%" ];
+      then
+        pw-metadata -n settings 0 clock.force-rate 0 > /dev/null
 
-        if [ "$currentRate" != "$pwrate" ];
+        break
+      fi
+      if [ "$currentRate" != "$pwrate" ];
+      then
+        if [[ " ''${allowedRates[*]}" =~ [[:space:]]''${currentRate}[[:space:]] ]];
         then
-          if [[ " ''${allowedRates[*]}" =~ [[:space:]]''${currentRate}[[:space:]] ]];
+          pw-metadata -n settings 0 clock.force-rate "$currentRate" > /dev/null
+          pwrate=$(pw-metadata -n settings | grep 'clock.force-rate' | cut -d "'" -f 4)
+
+        elif [ $(( currentRate % 48000 )) -eq 0 ];
+        then
+          if [[ " ''${allowedRates[*]}" =~ [[:space:]]192000[[:space:]] ]];
           then
-            pw-metadata -n settings 0 clock.force-rate "$currentRate" > /dev/null
-            pwrate=$(pw-metadata -n settings | grep 'clock.force-rate' | cut -d "'" -f 4)
+            pw-metadata -n settings 0 clock.force-rate 192000 > /dev/null
+            pwrate=''${currentRate}
 
-          elif [ $(( currentRate % 48000 )) -eq 0 ];
+          elif [[ " ''${allowedRates[*]}" =~ [[:space:]]96000[[:space:]] ]];
           then
-            if [[ " ''${allowedRates[*]}" =~ [[:space:]]192000[[:space:]] ]];
-            then
-              pw-metadata -n settings 0 clock.force-rate 192000 > /dev/null
-              pwrate=''${currentRate}
+            pw-metadata -n settings 0 clock.force-rate 96000 > /dev/null
+            pwrate=''${currentRate}
 
-            elif [[ " ''${allowedRates[*]}" =~ [[:space:]]96000[[:space:]] ]];
-            then
-              pw-metadata -n settings 0 clock.force-rate 96000 > /dev/null
-              pwrate=''${currentRate}
-
-            elif [[ " ''${allowedRates[*]}" =~ [[:space:]]48000[[:space:]] ]];
-            then
-              pw-metadata -n settings 0 clock.force-rate 48000 > /dev/null
-              pwrate=''${currentRate}
-
-            else
-              kill "$pid"
-              pw-metadata -n settings 0 clock.force-rate 44100 > /dev/null
-              pwrate=''${currentRate}
-
-            fi
+          elif [[ " ''${allowedRates[*]}" =~ [[:space:]]48000[[:space:]] ]];
+          then
+            pw-metadata -n settings 0 clock.force-rate 48000 > /dev/null
+            pwrate=''${currentRate}
 
           else
-            if [[ " ''${allowedRates[*]}" =~ [[:space:]]88200[[:space:]] ]];
-            then
-              pw-metadata -n settings 0 clock.force-rate 88200 > /dev/null
-              pwrate=''${currentRate}
+            kill "$pid"
+            pw-metadata -n settings 0 clock.force-rate 44100 > /dev/null
+            pwrate=''${currentRate}
 
-            else
-              pw-metadata -n settings 0 clock.force-rate 44100 > /dev/null
-              pwrate=''${currentRate}
-
-            fi
           fi
 
+        else
+          if [[ " ''${allowedRates[*]}" =~ [[:space:]]88200[[:space:]] ]];
+          then
+            pw-metadata -n settings 0 clock.force-rate 88200 > /dev/null
+            pwrate=''${currentRate}
+
+          else
+            pw-metadata -n settings 0 clock.force-rate 44100 > /dev/null
+            pwrate=''${currentRate}
+
+          fi
         fi
+
       fi
 
     done
